toggle between map and graph
