+++ src/screens/prediction/style.scss
@@ -12,6 +12,76 @@
     margin-bottom: 30px;
 }
 
+.selection-p {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+    width: 300px;
+    height: 40px;
+    background-color: #FFFFFF;
+    border-radius: 100px;
+    border: 1px solid #B7B7B7;
+    box-shadow: 0px 0px 6px rgba(0, 0, 0, 0.25);
+    margin-left: 62vw;
+    margin-top: 5vh;
+
+    .selected-option-p {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        background-color: #DFFBFF;
+        border-radius: 100px;
+        padding-left: 25px;
+        padding-right: 25px;
+
+        .selected-option-text-p {
+            font-weight: bold;
+            color: #020202;
+        }
+        img {
+            height: 20px;
+            width: 20px;
+        }
+    }
+    .selected-option-2-p {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        background-color: #F1E0FE;
+        border-radius: 100px;
+        padding-left: 15px;
+        padding-right: 15px;
+
+        .selected-option-text-p {
+            font-weight: bold;
+            color: #020202;
+        }
+        img {
+            height: 20px;
+            width: 20px;
+            margin-right: 10px;
+        }
+    }
+    .unselected-option-p {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        padding-left: 15px;
+        padding-right: 15px;
+        cursor: pointer;
+
+        .unselected-option-text-p {
+            color: #93949B;
+        }
+        img {
+            height: 20px;
+            width: 20px;
+            margin-right: 10px;
+        }
+    }
+}
+
 .histogram {
     padding: 25px;
     border-radius: 26px;
